[unreachable-code] Fix the false positive in python 3.8
Refs #9751
