+++ tests/functional/r/regression_02/regression_9751.py
@@ -1,0 +1,15 @@
+"""
+pylint 3.2.4 regression
+https://github.com/pylint-dev/pylint/issues/9751
+"""
+
+# pylint: disable=missing-function-docstring
+
+from typing import Any
+
+def repro() -> Any:
+    return 5
+
+def main():
+    x = repro() + 5
+    print(x)
