+++ tests/functional/u/used/used_before_assignment_py38.rc
@@ -1,0 +1,2 @@
+[testoptions]
+min_pyver=3.9
